autoc-vizzoo-core/requirements.txt
django==1.11.16
djangorestframework==3.9.0
cx_Oracle==6.3.1
django-cors-headers==2.4.0
requests==2.21.0
django-rest-swagger==2.2.*
drf-yasg==1.13.*
django-choices==1.7.*